wallet: ensure best block is up to date during sync
This commit fixes a case during the initial sync of the wallet where
the code stops too early to scan the chain and as a consequence
fails to reach the birthday block, causing the initial start to stop
with error.

The scan is done in a loop until reaches the "bestHeight", where at
that point the code only exits if the chain backend also reports it is
synchronized at tip. Problem is that since the "bestHeight" is
initialized before the loop it might contain an old value causing the
main loop to exit too early.

The solution is to refresh the bestHeight on every iteration and make
sure it is always called after isCurrent is called and might return
true.